+++ src/js/components/datepicker.js
@@ -168,8 +168,8 @@
 
             this.current  = this.element.val() ? moment(this.element.val(), this.options.format) : moment();
 
-            this.on("click", function(){
-                if (active!==$this) $this.pick(this.value);
+            this.on("click focus", function(){
+                if (active!==$this) $this.pick(this.value ? this.value:($this.options.minDate ? $this.options.minDate :''));
             }).on("change", function(){
 
                 if ($this.element.val() && !moment($this.element.val(), $this.options.format).isValid()) {