+++ src/js/core/tab.js
@@ -32,7 +32,10 @@
 
             var $this = this;
 
+            this.current = false;
+
             this.on("click.uikit.tab", this.options.target, function(e) {
+
                 e.preventDefault();
 
                 if ($this.switcher && $this.switcher.animating) {
@@ -42,7 +45,10 @@
                 var current = $this.find($this.options.target).not(this);
 
                 current.removeClass("uk-active").blur();
-                $this.trigger("change.uk.tab", [UI.$(this).addClass("uk-active")]);
+
+                $this.trigger("change.uk.tab", [UI.$(this).addClass("uk-active"), $this.current]);
+
+                $this.current = UI.$(this);
 
                 // Update ARIA
                 if (!$this.options.connect) {